+++ snowddl/parser/procedure.py
@@ -1,4 +1,4 @@
-from snowddl.blueprint import ProcedureBlueprint, Ident, SchemaObjectIdentWithArgs, NameWithType, DataType
+from snowddl.blueprint import ProcedureBlueprint, Ident, SchemaObjectIdentWithArgs, NameWithType, DataType, StageWithPath, build_schema_object_ident
 from snowddl.parser.abc_parser import AbstractParser, ParsedFile
 
 
@@ -18,7 +18,17 @@ procedure_json_schema = {
             }
         },
         "returns": {
-            "type": "string"
+            "anyOf": [
+                {
+                    "type": "string"
+                },
+                {
+                    "type": "object",
+                    "additionalProperties": {
+                        "type": "string"
+                    }
+                }
+            ]
         },
         "is_strict": {
             "type": "boolean"
@@ -26,11 +36,48 @@ procedure_json_schema = {
         "is_execute_as_caller": {
             "type": "boolean"
         },
+        "runtime_version": {
+            "type": "string"
+        },
+        "imports": {
+            "type": "array",
+            "items": {
+                "type": "object",
+                "properties": {
+                    "stage": {
+                        "type": "string"
+                    },
+                    "path": {
+                        "type": "string"
+                    }
+                },
+                "required": ["stage"],
+                "additionalProperties": False
+            },
+            "minItems": 1
+        },
+        "packages": {
+            "type": "array",
+            "items": {
+                "type": "string"
+            },
+            "minItems": 1
+        },
+        "handler": {
+            "type": "string"
+        },
         "comment": {
             "type": "string"
         }
     },
-    "required": ["body", "returns"],
+    "anyOf": [
+        {
+            "required": ["body", "returns"]
+        },
+        {
+            "required": ["handler", "returns"]
+        }
+    ],
     "additionalProperties": False
 }
 
@@ -41,17 +88,39 @@ class ProcedureParser(AbstractParser):
 
     def process_procedure(self, f: ParsedFile):
         arguments = [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params.get('arguments', {}).items()]
+
+        if isinstance(f.params['returns'], dict):
+            # Returns table
+            returns = [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params['returns'].items()]
+        else:
+            # Returns value
+            returns = DataType(f.params['returns'])
+
+        if f.params.get('imports'):
+            imports = [StageWithPath(stage_name=build_schema_object_ident(self.env_prefix, i['stage'], f.database, f.schema), path=i['path']) for i in f.params.get('imports')]
+        else:
+            imports = None
+
+        if f.params.get('packages'):
+            packages = f.params.get('packages')
+        else:
+            packages = None
+
         base_name = self.validate_name_with_args(f.path, arguments)
 
         bp = ProcedureBlueprint(
             full_name=SchemaObjectIdentWithArgs(self.env_prefix, f.database, f.schema, base_name, [a.type.base_type for a in arguments]),
             language=f.params.get('language', 'SQL'),
-            body=f.params['body'],
+            body=f.params.get('body'),
             arguments=arguments,
-            returns=DataType(f.params['returns']),
+            returns=returns,
             is_strict=f.params.get('is_strict', False),
             is_immutable=f.params.get('is_immutable', False),
             is_execute_as_caller=f.params.get('is_execute_as_caller', False),
+            runtime_version=f.params.get('runtime_version'),
+            imports=imports,
+            packages=packages,
+            handler=f.params.get('handler'),
             comment=f.params.get('comment'),
         )